requirements.txt
ruff
libusb1
numpy
hexdump
pycryptodome
tqdm>=4.14.0
pytest
pytest-mock
pytest-xdist
pytest-timeout
pytest-randomly
parameterized
cffi
pre-commit
scons>=4.4.0
flaky
termcolor